Update and rename menu.c to menu(수정).c
